[PBA-5408] FullScreenHelper added to support a full screen mode with RecyclerView
